chore: explore endpoint based solution
